Restructure project for scalability
Project split into 3 separate directories:
- crud
- models
- routers
