+++ question-service/.gitignore
@@ -1,4 +1,4 @@
 .venv
-__pycache__
+**/__pycache__
 
 .env
